+++ lib/src/models/drawer_menu.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class DrawerMenu {
+  DrawerMenu({@required this.child, @required this.onPressed});
+
+  final Widget child;
+  final VoidCallback onPressed;
+}