+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## [0.0.1] - TODO: Add release date.
+## 0.1.0 - 20/09/2019
 
-* TODO: Describe initial release.
+- Add HiddenDrawerMenu for the drawer that accept list of DrawerMenu
+- Add provider to listen the drawer menu state
+- Update environtment version to >=2.3.0 <3.0.0
+
+## 0.0.1 - 19/09/2019
+
+- Initial release hidden drawer